音调支持 1 位小数：滑块 0.1 半音步进（-16.0~+16.0），pitch 全链路改 float（widgets/controller/params/offline_config/offline_manager/window/change_key），支持 +12.5 等精确对齐

+++ gui/infer/controller.py
@@ -13,7 +13,7 @@ logger = logging.getLogger(__name__)
 
 @dataclass
 class ModelConfig:
-    pitch: int
+    pitch: float
     gender: float
     protect: float
     f0method: str
